#2_[feat] - 머지 후 리팩토링
